+++ hwbench/engines/stressng.py
@@ -27,24 +27,6 @@ class EngineModulePinnable(EngineModuleBase):
         return ""
 
 
-class EngineModuleQsort(EngineModulePinnable):
-    """This class implements the Qsort EngineModuleBase for StressNG"""
-
-    def __init__(self, engine: EngineBase, engine_module_name: str):
-        super().__init__(engine, engine_module_name)
-        self.engine_module_name = engine_module_name
-        self.add_module_parameter("qsort")
-
-    def run_cmd(self, p: BenchmarkParameters):
-        return StressNGQsort(self, p).run_cmd()
-
-    def run(self, p: BenchmarkParameters):
-        return StressNGQsort(self, p).run()
-
-    def fully_skipped_job(self, p) -> bool:
-        return StressNGQsort(self, p).fully_skipped_job()
-
-
 class EngineModuleMemrate(EngineModulePinnable):
     """This class implements the Memrate EngineModuleBase for StressNG"""
 
@@ -121,6 +103,7 @@ class Engine(EngineBase):
 
     def __init__(self):
         from .stressng_cpu import EngineModuleCpu
+        from .stressng_qsort import EngineModuleQsort
 
         super().__init__("stressng", "stress-ng")
         self.add_module(EngineModuleCpu(self, "cpu"))
@@ -258,17 +241,6 @@ class StressNG(ExternalBench):
             "effective_runtime": 0,
             "skipped": True,
         }
-
-
-class StressNGQsort(StressNG):
-    """The StressNG Qsort CPU stressor."""
-
-    def run_cmd(self) -> list[str]:
-        # Let's build the command line to run the tool
-        return super().run_cmd() + [
-            "--qsort",
-            str(self.parameters.get_engine_instances_count()),
-        ]
 
 
 class StressNGStream(StressNG):
